+++ qucs/diagrams/smithdiagram.h
@@ -24,7 +24,7 @@
 class SmithDiagram : public Diagram  {
 public: 
   SmithDiagram(int _cx=0, int _cy=0, bool ImpMode=true);
-  ~SmithDiagram();
+ ~SmithDiagram();
 
 
   Diagram* newOne();
@@ -32,7 +32,7 @@ public:
   static Element* info_y(QString&, char* &, bool getNewOne=false);
   int  calcDiagram();
   void calcLimits();
-  void calcCoordinate(double* &, double* &, double* &, int*, int*, Axis*);
+  void calcCoordinate(double* &, double* &, double* &, float*, float*, Axis*);
 };
 
 #endif
